+++ packages/ZbayLite/src/renderer/components/utils.js
@@ -1,5 +1,0 @@
-export const getZbayAddress = (zcashAddress) => `zbay.io/uri/${zcashAddress}`
-
-export default {
-  getZbayAddress
-}