+++ packages/ZbayLite/src/renderer/containers/widgets/channels/ChannelMenuAction.js
@@ -2,9 +2,10 @@ import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
 
 import ChannelMenuAction from '../../../components/widgets/channels/ChannelMenuAction'
+import { actionCreators } from '../../../store/handlers/modals'
 
 export const mapDispatchToProps = dispatch => bindActionCreators({
-  onInfo: () => console.warn('[ChannelMenuAction] onInfo not implemented'),
+  onInfo: actionCreators.openModal('channelInfo'),
   onMute: () => console.warn('[ChannelMenuAction] onMute not implemented'),
   onDelete: () => console.warn('[ChannelMenuAction] onDelete not implemented')
 }, dispatch)