+++ packages/ZbayLite/src/renderer/store/handlers/channel.test.js
@@ -83,6 +83,12 @@ describe('channel reducer', () => {
       const channel = channelSelectors.channel(store.getState())
       expect(channel).toMatchSnapshot()
     })
+
+    it('- setShareableUri', () => {
+      store.dispatch(actions.setShareableUri('zbay.io/channel/this-is-a-hash'))
+      const channel = channelSelectors.channel(store.getState())
+      expect(channel).toMatchSnapshot()
+    })
   })
 
   describe('handles epics', () => {