+++ packages/ZbayLite/src/renderer/store/handlers/importedChannel.test.js
@@ -5,6 +5,7 @@ jest.mock('../../zcash')
 import Immutable from 'immutable'
 
 import create from '../create'
+import { uriToChannel } from '../../zbay/channels'
 import importedChannelHandlers, { ImportedChannelState } from './importedChannel'
 import importedChannelSelectors from '../selectors/importedChannel'
 import channelsSelectors from '../selectors/channels'
@@ -32,36 +33,34 @@ describe('Imported channel reducer handles', () => {
     jest.clearAllMocks()
   })
 
-  const channelUri = 'zbay.io/channel/eJxNj8FugzAMhl8lyrmaOgoM9hw97WbAISHBDYmhTaq++9LDpEmW/oM/+f/8lAQrym95xciGZjFqIEInT9IHcwCXHYcdTxKmKWCMBc1c2AjeFf6zxfqOdlvSozunZcmqqjMn5bgaG+YWDq5J8UxL29fUYfPoVG586HGJ08X1bk2rqnqrF7Klc8I4BuPZ3OjtpE0UZUDMSBjA/dkJOMA4GBwKvglwTuwRQxQ3JX4GSGJIYkIFu+OPctRiKtpPaQ77ti/x/wHfBF+NObPrYAW3rc2s86HDpaFzp6v20F7TvXq0trtwUPwVt0j9ZqdWvl6/PI1yUA=='
+  const channelUri = 'zbay.io/channel/eJxNj0tuhDAQRK9ieR1FE36BXCAXyCq7BtoY3DRg9zBjj+bucZRNpJJKqkXVq4dmWFF/6E9k9ED6RcM4egwhZ0kwSICdZp7eGqxu6I4l3ttLXJZkiipJNCTFUIs0cErFRiZemq7iFut7a1K9+w6XMJbU0RpXU3TOLuzyyIhh8PMu88Z56MvOQWWBmv4w1GCBGUnBCTNBT6hkU0CkrgF9UJtR3z1E1Uc1ooEryWsudRgz9kPPp/ulz/b/wF77vRhSEmphBTrWerLptL6s+dLaojntbvlW3BvXluKNvIcjcHe4sdHP5w8SWGnS'
 
   describe('actions', () => {
-    it('handles decodeChannel', async () => {
-      await store.dispatch(importedChannelHandlers.actions.decodeChannel(channelUri))
+    it('handles setData', async () => {
+      const channel = await uriToChannel(channelUri)
+      await store.dispatch(importedChannelHandlers.actions.setData(channel))
 
-      const decoding = importedChannelSelectors.decoding(store.getState())
-      const errors = importedChannelSelectors.errors(store.getState())
       const data = importedChannelSelectors.data(store.getState())
-      expect(decoding).toBeFalsy()
-      expect(errors).toBeFalsy()
       expect(data).toMatchSnapshot()
     })
 
-    it('handles failing decodeChannel', async () => {
-      expect.assertions(3)
-      try {
-        await store.dispatch(importedChannelHandlers.actions.decodeChannel('incorrect-uri'))
-      } catch (err) {
-        const decoding = importedChannelSelectors.decoding(store.getState())
-        const errors = importedChannelSelectors.errors(store.getState())
-        const data = importedChannelSelectors.data(store.getState())
-        expect(decoding).toBeFalsy()
-        expect(data).toBeNull()
-        expect(errors).toMatchSnapshot()
-      }
+    it('handles setDecoding', async () => {
+      await store.dispatch(importedChannelHandlers.actions.setDecoding(true))
+
+      const decoding = importedChannelSelectors.decoding(store.getState())
+      expect(decoding).toBeTruthy()
+    })
+
+    it('handles setDecodingError', async () => {
+      await store.dispatch(importedChannelHandlers.actions.setDecodingError(new Error('this is a test error')))
+
+      const errors = importedChannelSelectors.errors(store.getState())
+      expect(errors).toMatchSnapshot()
     })
 
     it('handles clear', async () => {
-      await store.dispatch(importedChannelHandlers.actions.decodeChannel(channelUri))
+      const channel = await uriToChannel(channelUri)
+      await store.dispatch(importedChannelHandlers.actions.setData(channel))
       store.dispatch(importedChannelHandlers.actions.clear())
 
       const decoding = importedChannelSelectors.decoding(store.getState())
@@ -82,7 +81,7 @@ describe('Imported channel reducer handles', () => {
             importChannel: importChannelMock
           }
         }))
-        await store.dispatch(importedChannelHandlers.actions.decodeChannel(channelUri))
+        await store.dispatch(importedChannelHandlers.epics.decodeChannel(channelUri))
 
         await store.dispatch(importedChannelHandlers.epics.importChannel())
 
@@ -91,7 +90,7 @@ describe('Imported channel reducer handles', () => {
 
       it('reloads channels', async () => {
         mock.setArchive(createArchive())
-        await store.dispatch(importedChannelHandlers.actions.decodeChannel(channelUri))
+        await store.dispatch(importedChannelHandlers.epics.decodeChannel(channelUri))
 
         await store.dispatch(importedChannelHandlers.epics.importChannel())
 
@@ -103,7 +102,7 @@ describe('Imported channel reducer handles', () => {
 
       it('dispatches notification on success', async () => {
         mock.setArchive(createArchive())
-        await store.dispatch(importedChannelHandlers.actions.decodeChannel(channelUri))
+        await store.dispatch(importedChannelHandlers.epics.decodeChannel(channelUri))
 
         await store.dispatch(importedChannelHandlers.epics.importChannel())
 
@@ -119,7 +118,7 @@ describe('Imported channel reducer handles', () => {
             importChannel: jest.fn(() => Promise.reject(new Error('test error')))
           }
         }))
-        await store.dispatch(importedChannelHandlers.actions.decodeChannel(channelUri))
+        await store.dispatch(importedChannelHandlers.epics.decodeChannel(channelUri))
 
         await store.dispatch(importedChannelHandlers.epics.importChannel())